Long is already coerced in the constructor, so we need not re-coerce and do a double boxing/unboxing here.

+++ src/main/java/org/junit/contrib/truth/subjects/IntegerSubject.java
@@ -89,8 +89,7 @@ public class IntegerSubject extends Subject<IntegerSubject, Long> {
         fail("is equal to", other);
       }
     } else {
-      // Coerce to a long.
-      if (!Long.valueOf(getSubject().longValue()).equals(other)) {
+      if (!getSubject().equals(other)) {
         fail("is equal to", other);
       }
     }
